compile fixes
git-svn-id: https://open.syn3.nl/syn3/svndav/default/trunk/projects/synapse@305 99606f1f-bbb2-4562-8e6b-387ddef1dc61

+++ modules/timer.module/module.cpp
@@ -69,7 +69,7 @@ SYNAPSE_REGISTER(timer_Set)
 		}
 		struct timespec abstime;
 		abstime.tv_sec=((int)msg["time"]);
-		abstime.tv_nsec=((long double)msg["time"] - (int)msg["time"]  )*1000000000;
+		abstime.tv_nsec=(long int)((long double)msg["time"] - (int)msg["time"]  )*1000000000;
 		clock_nanosleep(CLOCK_REALTIME, TIMER_ABSTIME, &abstime ,NULL);
 
 		Cmsg out(msg);
@@ -104,7 +104,7 @@ SYNAPSE_REGISTER(timer_Set)
 		{
 			struct timespec reltime;
 			reltime.tv_sec=((int)msg["seconds"]);
-			reltime.tv_nsec=((long double)msg["seconds"] - (int)msg["seconds"]  )*1000000000;
+			reltime.tv_nsec=(long int)((long double)msg["seconds"] - (int)msg["seconds"]  )*1000000000;
 			clock_nanosleep(CLOCK_REALTIME, 0, &reltime ,NULL);
 			out.send();
 			count++;
